requirements.txt
Django==4.2.11
Pillow>=10.0.0
django-crispy-forms==1.14.0
djangorestframework==3.15.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
drf-nested-routers>=0.64.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
black>=23.0.0
flake8>=6.1.0
isort>=5.13.0
# Production packages
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0
django-redis>=5.4.0
python-decouple>=3.8
python-dotenv>=1.0.0
